% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_lat_band.R
\name{bandit}
\alias{bandit}
\title{Rarefy localities within latitudinal bands}
\usage{
bandit(
  dat,
  xy,
  width,
  iter,
  nSite,
  centr = FALSE,
  absLat = FALSE,
  output = "locs"
)
}
\arguments{
\item{dat}{A \code{data.frame} or \code{matrix} containing the
coordinate columns \code{xy} and any associated variables, e.g. taxon names.}

\item{xy}{A vector of two elements, specifying the name or numeric position
of the columns containing longitude and latitude coordinates, respectively.}

\item{width}{A numeric value for the resolution of latitude bins, in degrees.}

\item{iter}{The number of times to subsample localities within \strong{each}
latitudinal band.}

\item{nSite}{The quota of unique locations to include in each subsample.}

\item{centr}{Logical: should a bin center on and cover the equator
(\code{TRUE}) or should the equator mark the boundary between the
lowest-latitude northern and southern bins (\code{FALSE}, default)?}

\item{absLat}{Logical: should only the absolute values of latitude be
evaluated? If \code{absLat = TRUE}, \code{centr} argument is ignored.}

\item{output}{Whether the returned data should be a two-column matrix of
subsample site coordinates (\code{output = 'locs'}) or the subset of rows
from \code{dat} associated with those coordinates (\code{output = 'full'}).}
}
\description{
\code{bandit} subsamples spatial point data to a specified number of sites
within bins of equal latitude
}
\references{
\insertRef{Marcot2016}{divvy}
}
