% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_eco_variables.R
\name{rangeSizer}
\alias{rangeSizer}
\title{Calculate common metrics of spatial distribution}
\usage{
rangeSizer(coords, crs)
}
\arguments{
\item{coords}{2-column \code{data.frame} or \code{matrix} containing
x- and y-coordinates, respectively (e.g. longitude and latitude).}

\item{crs}{Coordinate reference system as a GDAL text string, EPSG code,
or object of class \code{crs}. Default is latitude-longitude (\code{EPSG:4326}).}
}
\value{
A 1-row \code{data.frame} listing: count of spatial occurrences (including
duplicates) and their centroid coordinates, latitudinal range (degrees),
great circle distance (km), and summed minimum spanning tree length (km).
}
\description{
Coordinates and their distances are calculated with respect to
the original coordinate reference system if supplied, and otherwise
calculated as spherical distances between points assumed to be given
in degrees latitude-longitude. Projected coordinates will be transformed
to calculate latitudinal range in degrees.
}
