% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_eco_variables.R
\name{sampMeta}
\alias{sampMeta}
\title{Calculate basic spatial coverage and diversity metrics}
\usage{
sampMeta(
  dat,
  taxVar,
  xy,
  collections = NULL,
  quotaQ = NULL,
  quotaN = NULL,
  omitDom = FALSE
)
}
\arguments{
\item{dat}{A \code{data.frame} or \code{matrix} containing taxon names,
coordinates, and any associated variables.}

\item{taxVar}{The name or numeric position of the column containing
taxonomic identifications.}

\item{xy}{A vector of two elements, specifying the name or numeric position
of the columns containing longitude and latitude coordinates, respectively.}

\item{collections}{The name or numeric position of the column containing
unique collection IDs, e.g. 'collection_no' in PBDB data downloads.}

\item{quotaQ}{A numeric value for the coverage (quorum) level at which to
perform coverage-based rarefaction (shareholder quorum subsampling).}

\item{quotaN}{A numeric value for the quota of taxon occurrences to subsample
in classical rarefaction.}

\item{omitDom}{If \code{omitDom = TRUE} and \code{quotaQ} or \code{quotaN}
is supplied, remove the most common taxon prior to rarefaction.}
}
\description{
Summarise the geographic scope and position of occurrence data, and
optionally estimate richness and evenness
}
\details{
\code{sampMeta} compiles a variety of metadata about a (sub)sample,
before or after spatial subsampling. The function counts the number
of unique spatial sites, collections (if requested), and taxa, and
calculates the centroid coordinates, latitudinal range (degrees),
great circle distance (km), and summed minimum spanning tree length (km)
for occurrences.

If \code{quotaQ} is supplied, \code{sampMeta} rarefies richness at the
given coverage value and returns the point estimate of richness (Hill number 0)
and its 95\% confidence interval as well as estimates of evenness (Pielou's J)
and sample coverage (Good's \emph{u}). If \code{quotaN} is supplied,
\code{sampMeta} rarefies richness to the given number of occurrence counts
and returns the point estimate of richness and its 95\% confidence interval.

Coverage-based and classical rarefaction are both calculated with
\code{iNEXT::estimateD} internally. For details, such as how diversity
is extrapolated if a sample is insufficient to achieve a specified
rarefaction level, consult Chao and Jost (2012) and Hsieh \emph{et al.} (2016).
}
\references{
\insertRef{Chao2012}{divDivvy}

\insertRef{Hsieh2016}{divDivvy}
}
